kservice: fix almost all clang warnings
except in lex/yacc generated code, some -Wmissing-prototypes cases (including in Qt macros),
moc calling deprecated methods, and two if(false) in substring support (for which I emailed aseigo).

My flags:
-Weverything -Wno-c++98-compat -Wno-c++98-compat-pedantic -Wno-exit-time-destructors -Wno-padded -Wno-sign-conversion -Wno-documentation -Wno-undefined-reinterpret-cast -Wno-shorten-64-to-32 -Wno-conversion -Wno-gnu-zero-variadic-macro-arguments -Wno-global-constructors -Wno-shadow -Werror=format-security -Werror=return-type -Wno-variadic-macros -Wmissing-include-dirs -Wno-used-but-marked-unused -Wall -Wextra -Wcast-align -Wchar-subscripts -Wformat-security -Wno-long-long -Wpointer-arith -Wundef -Wnon-virtual-dtor -Woverloaded-virtual -pedantic -std=c++0x -fdeprecated-macro -fdebug-compilation-dir

REVIEW: 124772

+++ src/kbuildsycoca/vfolder_menu.cpp
@@ -430,7 +430,7 @@ VFolderMenu::absoluteDir(const QString &_dir, const QString &baseDir, bool keepR
 static void tagBaseDir(QDomDocument &doc, const QString &tag, const QString &dir)
 {
     QDomNodeList mergeFileList = doc.elementsByTagName(tag);
-    for (int i = 0; i < (int)mergeFileList.count(); i++) {
+    for (int i = 0; i < mergeFileList.count(); i++) {
         QDomAttr attr = doc.createAttribute(QStringLiteral("__BaseDir"));
         attr.setValue(dir);
         mergeFileList.item(i).toElement().setAttributeNode(attr);
@@ -440,7 +440,7 @@ static void tagBaseDir(QDomDocument &doc, const QString &tag, const QString &dir
 static void tagBasePath(QDomDocument &doc, const QString &tag, const QString &path)
 {
     QDomNodeList mergeFileList = doc.elementsByTagName(tag);
-    for (int i = 0; i < (int)mergeFileList.count(); i++) {
+    for (int i = 0; i < mergeFileList.count(); i++) {
         QDomAttr attr = doc.createAttribute(QStringLiteral("__BasePath"));
         attr.setValue(path);
         mergeFileList.item(i).toElement().setAttributeNode(attr);
